fix: move DOMAIN contant to const.py
