+++ .pre-commit-config.yaml
@@ -1,11 +1,4 @@
 repos:
-  - repo: local
-    hooks:
-      - id: manifest
-        name: Update manifest.json
-        entry: bin/update_manifest
-        language: script
-        files: ^(custom_components/.+/const\.py|requirements\.txt)$
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v4.4.0
     hooks:
