+++ custom_components/soma_connect/const.py
@@ -1,0 +1,3 @@
+"""Constants for SOMA Connect."""
+
+DOMAIN = "soma_connect"
