+++ custom_components/soma_connect/cover.py
@@ -17,7 +17,7 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import DOMAIN
+from .const import DOMAIN
 from .coordinator import SomaConnectUpdateCoordinator
 from .entity import SomaConnectEntity
