+++ custom_components/soma_connect/manifest.json
@@ -1,7 +1,7 @@
 {
   "domain": "soma_connect",
   "name": "SOMA Connect",
-  "version": "2022.12.2",
+  "version": "2022.12.3",
   "config_flow": true,
   "documentation": "https://github.com/Djelibeybi/hass-soma-connect",
   "issue_tracker": "https://github.com/Djelibeybi/hass-soma-connect/issues",
