+++ custom_components/soma_connect/sensor.py
@@ -14,7 +14,7 @@ from homeassistant.core import HomeAssistant, callback
 from homeassistant.helpers.entity import EntityCategory
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import DOMAIN
+from .const import DOMAIN
 from .coordinator import SomaConnectUpdateCoordinator
 from .entity import SomaConnectEntity
